Remove unused cache module and fetch cards from DB
